Revert "Refactor test driver"
