+++ .travis.yml
@@ -3,18 +3,11 @@ python:
   - "3.5"
 # command to install dependencies
 install: "pip install -r requirements.txt"
-
 # command to run tests
 script: 
   - py.test -v --color=yes
-  - bash etc/deploy_docs.sh
 branches:
   only:
   - master
   - development
-notifications:
-  email:
-    recipients:
-      - achyudhk@gmail.com
-    on_success: change
-    on_failure: always
+
