+++ docs/analysis.author.graph.html
@@ -1,0 +1,149 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+
+
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+    
+    <title>analysis.author.graph package &#8212; MailingListParser  documentation</title>
+    
+    <link rel="stylesheet" href="_static/alabaster.css" type="text/css" />
+    <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
+    
+    <script type="text/javascript">
+      var DOCUMENTATION_OPTIONS = {
+        URL_ROOT:    './',
+        VERSION:     '',
+        COLLAPSE_INDEX: false,
+        FILE_SUFFIX: '.html',
+        HAS_SOURCE:  true,
+        SOURCELINK_SUFFIX: '.txt'
+      };
+    </script>
+    <script type="text/javascript" src="_static/jquery.js"></script>
+    <script type="text/javascript" src="_static/underscore.js"></script>
+    <script type="text/javascript" src="_static/doctools.js"></script>
+    <link rel="index" title="Index" href="genindex.html" />
+    <link rel="search" title="Search" href="search.html" />
+   
+  <link rel="stylesheet" href="_static/custom.css" type="text/css" />
+  
+  <meta name="viewport" content="width=device-width, initial-scale=0.9, maximum-scale=0.9" />
+
+  </head>
+  <body role="document">
+  
+
+    <div class="document">
+      <div class="documentwrapper">
+        <div class="bodywrapper">
+          <div class="body" role="main">
+            
+  <div class="section" id="analysis-author-graph-package">
+<h1>analysis.author.graph package<a class="headerlink" href="#analysis-author-graph-package" title="Permalink to this headline">¶</a></h1>
+<div class="section" id="submodules">
+<h2>Submodules<a class="headerlink" href="#submodules" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="analysis-author-graph-author-interaction-module">
+<h2>analysis.author.graph.author_interaction module<a class="headerlink" href="#analysis-author-graph-author-interaction-module" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="module-analysis.author.graph.bipartite">
+<span id="analysis-author-graph-bipartite-module"></span><h2>analysis.author.graph.bipartite module<a class="headerlink" href="#module-analysis.author.graph.bipartite" title="Permalink to this headline">¶</a></h2>
+<p>This module is used to generate bipartite graph among all the users and messages in the mailing list such that all
+the users are on one side and all the messages will be on another. A directed edge would be drawn from author to the
+message sent by the author. A directed edge would be drawn from message to all the users who are in To and CC fields.
+A projection of this bipartite graph is then generated.</p>
+<dl class="function">
+<dt id="analysis.author.graph.bipartite.msg_author_bipartite_graph">
+<code class="descclassname">analysis.author.graph.bipartite.</code><code class="descname">msg_author_bipartite_graph</code><span class="sig-paren">(</span><em>threadwise=False</em>, <em>ignore_lat=True</em>, <em>time_limit=None</em><span class="sig-paren">)</span><a class="headerlink" href="#analysis.author.graph.bipartite.msg_author_bipartite_graph" title="Permalink to this definition">¶</a></dt>
+<dd><table class="docutils field-list" frame="void" rules="none">
+<col class="field-name" />
+<col class="field-body" />
+<tbody valign="top">
+<tr class="field-odd field"><th class="field-name">Parameters:</th><td class="field-body"><ul class="first simple">
+<li><strong>threadwise</strong> &#8211; If true, a bipartite graph is generated for each thread individually.</li>
+<li><strong>ignore_lat</strong> &#8211; If true, then messages that belong to threads that have only a single author are ignored.</li>
+<li><strong>time_limit</strong> &#8211; Time limit can be specified here in the form of a timestamp in one of the identifiable formats
+and all messages that have arrived after this timestamp will be ignored.</li>
+</ul>
+</td>
+</tr>
+<tr class="field-even field"><th class="field-name">Returns:</th><td class="field-body"><p class="first last"></p>
+</td>
+</tr>
+</tbody>
+</table>
+</dd></dl>
+
+</div>
+<div class="section" id="analysis-author-graph-generate-module">
+<h2>analysis.author.graph.generate module<a class="headerlink" href="#analysis-author-graph-generate-module" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="module-analysis.author.graph">
+<span id="module-contents"></span><h2>Module contents<a class="headerlink" href="#module-analysis.author.graph" title="Permalink to this headline">¶</a></h2>
+</div>
+</div>
+
+
+          </div>
+        </div>
+      </div>
+      <div class="sphinxsidebar" role="navigation" aria-label="main navigation">
+        <div class="sphinxsidebarwrapper">
+  <h3><a href="index.html">Table Of Contents</a></h3>
+  <ul>
+<li><a class="reference internal" href="#">analysis.author.graph package</a><ul>
+<li><a class="reference internal" href="#submodules">Submodules</a></li>
+<li><a class="reference internal" href="#analysis-author-graph-author-interaction-module">analysis.author.graph.author_interaction module</a></li>
+<li><a class="reference internal" href="#module-analysis.author.graph.bipartite">analysis.author.graph.bipartite module</a></li>
+<li><a class="reference internal" href="#analysis-author-graph-generate-module">analysis.author.graph.generate module</a></li>
+<li><a class="reference internal" href="#module-analysis.author.graph">Module contents</a></li>
+</ul>
+</li>
+</ul>
+<div class="relations">
+<h3>Related Topics</h3>
+<ul>
+  <li><a href="index.html">Documentation overview</a><ul>
+  </ul></li>
+</ul>
+</div>
+  <div role="note" aria-label="source link">
+    <h3>This Page</h3>
+    <ul class="this-page-menu">
+      <li><a href="_sources/analysis.author.graph.rst.txt"
+            rel="nofollow">Show Source</a></li>
+    </ul>
+   </div>
+<div id="searchbox" style="display: none" role="search">
+  <h3>Quick search</h3>
+    <form class="search" action="search.html" method="get">
+      <div><input type="text" name="q" /></div>
+      <div><input type="submit" value="Go" /></div>
+      <input type="hidden" name="check_keywords" value="yes" />
+      <input type="hidden" name="area" value="default" />
+    </form>
+</div>
+<script type="text/javascript">$('#searchbox').show(0);</script>
+        </div>
+      </div>
+      <div class="clearer"></div>
+    </div>
+    <div class="footer">
+      &copy;2017, Prasad Talasila.
+      
+      |
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 1.5.6</a>
+      &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.8</a>
+      
+      |
+      <a href="_sources/analysis.author.graph.rst.txt"
+          rel="nofollow">Page source</a>
+    </div>
+
+    
+
+    
+  </body>
+</html>
