+++ docs/analysis.thread.graph.html
@@ -1,0 +1,133 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+
+
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+    
+    <title>analysis.thread.graph package &#8212; MailingListParser  documentation</title>
+    
+    <link rel="stylesheet" href="_static/alabaster.css" type="text/css" />
+    <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
+    
+    <script type="text/javascript">
+      var DOCUMENTATION_OPTIONS = {
+        URL_ROOT:    './',
+        VERSION:     '',
+        COLLAPSE_INDEX: false,
+        FILE_SUFFIX: '.html',
+        HAS_SOURCE:  true,
+        SOURCELINK_SUFFIX: '.txt'
+      };
+    </script>
+    <script type="text/javascript" src="_static/jquery.js"></script>
+    <script type="text/javascript" src="_static/underscore.js"></script>
+    <script type="text/javascript" src="_static/doctools.js"></script>
+    <link rel="index" title="Index" href="genindex.html" />
+    <link rel="search" title="Search" href="search.html" />
+   
+  <link rel="stylesheet" href="_static/custom.css" type="text/css" />
+  
+  <meta name="viewport" content="width=device-width, initial-scale=0.9, maximum-scale=0.9" />
+
+  </head>
+  <body role="document">
+  
+
+    <div class="document">
+      <div class="documentwrapper">
+        <div class="bodywrapper">
+          <div class="body" role="main">
+            
+  <div class="section" id="analysis-thread-graph-package">
+<h1>analysis.thread.graph package<a class="headerlink" href="#analysis-thread-graph-package" title="Permalink to this headline">¶</a></h1>
+<div class="section" id="submodules">
+<h2>Submodules<a class="headerlink" href="#submodules" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="module-analysis.thread.graph.edge_list">
+<span id="analysis-thread-graph-edge-list-module"></span><h2>analysis.thread.graph.edge_list module<a class="headerlink" href="#module-analysis.thread.graph.edge_list" title="Permalink to this headline">¶</a></h2>
+<dl class="function">
+<dt id="analysis.thread.graph.edge_list.generate_edge_list">
+<code class="descclassname">analysis.thread.graph.edge_list.</code><code class="descname">generate_edge_list</code><span class="sig-paren">(</span><em>nodelist_filename='graph_nodes.csv'</em>, <em>edgelist_filename='graph_edges.csv'</em>, <em>json_filename='clean_data.json'</em><span class="sig-paren">)</span><a class="headerlink" href="#analysis.thread.graph.edge_list.generate_edge_list" title="Permalink to this definition">¶</a></dt>
+<dd><p>This function generates a list of nodes and edges in the graphs from the JSON file and saves it as a CSV file.
+:param ref_toggle: If True, References attribute is used to make edges and if False, In-Reply-To is used.</p>
+</dd></dl>
+
+<dl class="function">
+<dt id="analysis.thread.graph.edge_list.generate_node_labels">
+<code class="descclassname">analysis.thread.graph.edge_list.</code><code class="descname">generate_node_labels</code><span class="sig-paren">(</span><em>nodelist_filename='graph_nodes.txt'</em>, <em>edgelist_filename='graph_edges.txt'</em>, <em>json_filename='clean_data.json'</em><span class="sig-paren">)</span><a class="headerlink" href="#analysis.thread.graph.edge_list.generate_node_labels" title="Permalink to this definition">¶</a></dt>
+<dd><p>This function generates a list of nodes and edges in the graphs from the JSON file and saves it as a CSV file.
+:param ref_toggle: If True, References attribute is used to make edges and if False, In-Reply-To is used.</p>
+</dd></dl>
+
+</div>
+<div class="section" id="analysis-thread-graph-generate-module">
+<h2>analysis.thread.graph.generate module<a class="headerlink" href="#analysis-thread-graph-generate-module" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="module-analysis.thread.graph">
+<span id="module-contents"></span><h2>Module contents<a class="headerlink" href="#module-analysis.thread.graph" title="Permalink to this headline">¶</a></h2>
+</div>
+</div>
+
+
+          </div>
+        </div>
+      </div>
+      <div class="sphinxsidebar" role="navigation" aria-label="main navigation">
+        <div class="sphinxsidebarwrapper">
+  <h3><a href="index.html">Table Of Contents</a></h3>
+  <ul>
+<li><a class="reference internal" href="#">analysis.thread.graph package</a><ul>
+<li><a class="reference internal" href="#submodules">Submodules</a></li>
+<li><a class="reference internal" href="#module-analysis.thread.graph.edge_list">analysis.thread.graph.edge_list module</a></li>
+<li><a class="reference internal" href="#analysis-thread-graph-generate-module">analysis.thread.graph.generate module</a></li>
+<li><a class="reference internal" href="#module-analysis.thread.graph">Module contents</a></li>
+</ul>
+</li>
+</ul>
+<div class="relations">
+<h3>Related Topics</h3>
+<ul>
+  <li><a href="index.html">Documentation overview</a><ul>
+  </ul></li>
+</ul>
+</div>
+  <div role="note" aria-label="source link">
+    <h3>This Page</h3>
+    <ul class="this-page-menu">
+      <li><a href="_sources/analysis.thread.graph.rst.txt"
+            rel="nofollow">Show Source</a></li>
+    </ul>
+   </div>
+<div id="searchbox" style="display: none" role="search">
+  <h3>Quick search</h3>
+    <form class="search" action="search.html" method="get">
+      <div><input type="text" name="q" /></div>
+      <div><input type="submit" value="Go" /></div>
+      <input type="hidden" name="check_keywords" value="yes" />
+      <input type="hidden" name="area" value="default" />
+    </form>
+</div>
+<script type="text/javascript">$('#searchbox').show(0);</script>
+        </div>
+      </div>
+      <div class="clearer"></div>
+    </div>
+    <div class="footer">
+      &copy;2017, Prasad Talasila.
+      
+      |
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 1.5.6</a>
+      &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.8</a>
+      
+      |
+      <a href="_sources/analysis.thread.graph.rst.txt"
+          rel="nofollow">Page source</a>
+    </div>
+
+    
+
+    
+  </body>
+</html>
