+++ docs/input.imap.html
@@ -1,0 +1,146 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+
+
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+    
+    <title>input.imap package &#8212; MailingListParser  documentation</title>
+    
+    <link rel="stylesheet" href="_static/alabaster.css" type="text/css" />
+    <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
+    
+    <script type="text/javascript">
+      var DOCUMENTATION_OPTIONS = {
+        URL_ROOT:    './',
+        VERSION:     '',
+        COLLAPSE_INDEX: false,
+        FILE_SUFFIX: '.html',
+        HAS_SOURCE:  true,
+        SOURCELINK_SUFFIX: '.txt'
+      };
+    </script>
+    <script type="text/javascript" src="_static/jquery.js"></script>
+    <script type="text/javascript" src="_static/underscore.js"></script>
+    <script type="text/javascript" src="_static/doctools.js"></script>
+    <link rel="index" title="Index" href="genindex.html" />
+    <link rel="search" title="Search" href="search.html" />
+   
+  <link rel="stylesheet" href="_static/custom.css" type="text/css" />
+  
+  <meta name="viewport" content="width=device-width, initial-scale=0.9, maximum-scale=0.9" />
+
+  </head>
+  <body role="document">
+  
+
+    <div class="document">
+      <div class="documentwrapper">
+        <div class="bodywrapper">
+          <div class="body" role="main">
+            
+  <div class="section" id="input-imap-package">
+<h1>input.imap package<a class="headerlink" href="#input-imap-package" title="Permalink to this headline">¶</a></h1>
+<div class="section" id="submodules">
+<h2>Submodules<a class="headerlink" href="#submodules" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="module-input.imap.connection">
+<span id="input-imap-connection-module"></span><h2>input.imap.connection module<a class="headerlink" href="#module-input.imap.connection" title="Permalink to this headline">¶</a></h2>
+<dl class="function">
+<dt id="input.imap.connection.open_connection">
+<code class="descclassname">input.imap.connection.</code><code class="descname">open_connection</code><span class="sig-paren">(</span><em>verbose=False</em><span class="sig-paren">)</span><a class="headerlink" href="#input.imap.connection.open_connection" title="Permalink to this definition">¶</a></dt>
+<dd><p>Function to establish to mail server and login to user account using IMAP4 protocol.
+Returns the connection object used in establishing IMAP connection to mail server.
+Takes as argument boolen value which is used to determine whether function works in
+verbose mode or not.</p>
+</dd></dl>
+
+</div>
+<div class="section" id="module-input.imap.header">
+<span id="input-imap-header-module"></span><h2>input.imap.header module<a class="headerlink" href="#module-input.imap.header" title="Permalink to this headline">¶</a></h2>
+<dl class="function">
+<dt id="input.imap.header.get_mail_header">
+<code class="descclassname">input.imap.header.</code><code class="descname">get_mail_header</code><span class="sig-paren">(</span><em>to_get</em>, <em>range_=True</em><span class="sig-paren">)</span><a class="headerlink" href="#input.imap.header.get_mail_header" title="Permalink to this definition">¶</a></dt>
+<dd><p>This function fetches the emails from the IMAP server as per the parameters passed.
+:param to_get: List of UIDs of the mails to get. Default value is 2000.
+:param <a href="#id1"><span class="problematic" id="id2">range_</span></a>: If true, fetches all emails from the first element in to_get, till the newest mail.
+If false, fetches only the emails with the UIDs present in to_get. Default value is true.</p>
+</dd></dl>
+
+<dl class="function">
+<dt id="input.imap.header.init_uid_map">
+<code class="descclassname">input.imap.header.</code><code class="descname">init_uid_map</code><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#input.imap.header.init_uid_map" title="Permalink to this definition">¶</a></dt>
+<dd><p>To ensure that references are correctly recorded in the JSON file such that there are no references to mails that
+do not exist and to ease the processing of headers, a map with the string in the Message-Id field of the header to
+the UID of the mail is required. This function reads the header.json file and adds required entries to the map.
+:return: A map with the string in the Message-Id field of the header to the UID of the mail</p>
+</dd></dl>
+
+</div>
+<div class="section" id="module-input.imap">
+<span id="module-contents"></span><h2>Module contents<a class="headerlink" href="#module-input.imap" title="Permalink to this headline">¶</a></h2>
+</div>
+</div>
+
+
+          </div>
+        </div>
+      </div>
+      <div class="sphinxsidebar" role="navigation" aria-label="main navigation">
+        <div class="sphinxsidebarwrapper">
+  <h3><a href="index.html">Table Of Contents</a></h3>
+  <ul>
+<li><a class="reference internal" href="#">input.imap package</a><ul>
+<li><a class="reference internal" href="#submodules">Submodules</a></li>
+<li><a class="reference internal" href="#module-input.imap.connection">input.imap.connection module</a></li>
+<li><a class="reference internal" href="#module-input.imap.header">input.imap.header module</a></li>
+<li><a class="reference internal" href="#module-input.imap">Module contents</a></li>
+</ul>
+</li>
+</ul>
+<div class="relations">
+<h3>Related Topics</h3>
+<ul>
+  <li><a href="index.html">Documentation overview</a><ul>
+  </ul></li>
+</ul>
+</div>
+  <div role="note" aria-label="source link">
+    <h3>This Page</h3>
+    <ul class="this-page-menu">
+      <li><a href="_sources/input.imap.rst.txt"
+            rel="nofollow">Show Source</a></li>
+    </ul>
+   </div>
+<div id="searchbox" style="display: none" role="search">
+  <h3>Quick search</h3>
+    <form class="search" action="search.html" method="get">
+      <div><input type="text" name="q" /></div>
+      <div><input type="submit" value="Go" /></div>
+      <input type="hidden" name="check_keywords" value="yes" />
+      <input type="hidden" name="area" value="default" />
+    </form>
+</div>
+<script type="text/javascript">$('#searchbox').show(0);</script>
+        </div>
+      </div>
+      <div class="clearer"></div>
+    </div>
+    <div class="footer">
+      &copy;2017, Prasad Talasila.
+      
+      |
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 1.5.6</a>
+      &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.8</a>
+      
+      |
+      <a href="_sources/input.imap.rst.txt"
+          rel="nofollow">Page source</a>
+    </div>
+
+    
+
+    
+  </body>
+</html>
