+++ docs/input.mbox.html
@@ -1,0 +1,134 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+
+
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+    
+    <title>input.mbox package &#8212; MailingListParser  documentation</title>
+    
+    <link rel="stylesheet" href="_static/alabaster.css" type="text/css" />
+    <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
+    
+    <script type="text/javascript">
+      var DOCUMENTATION_OPTIONS = {
+        URL_ROOT:    './',
+        VERSION:     '',
+        COLLAPSE_INDEX: false,
+        FILE_SUFFIX: '.html',
+        HAS_SOURCE:  true,
+        SOURCELINK_SUFFIX: '.txt'
+      };
+    </script>
+    <script type="text/javascript" src="_static/jquery.js"></script>
+    <script type="text/javascript" src="_static/underscore.js"></script>
+    <script type="text/javascript" src="_static/doctools.js"></script>
+    <link rel="index" title="Index" href="genindex.html" />
+    <link rel="search" title="Search" href="search.html" />
+   
+  <link rel="stylesheet" href="_static/custom.css" type="text/css" />
+  
+  <meta name="viewport" content="width=device-width, initial-scale=0.9, maximum-scale=0.9" />
+
+  </head>
+  <body role="document">
+  
+
+    <div class="document">
+      <div class="documentwrapper">
+        <div class="bodywrapper">
+          <div class="body" role="main">
+            
+  <div class="section" id="input-mbox-package">
+<h1>input.mbox package<a class="headerlink" href="#input-mbox-package" title="Permalink to this headline">¶</a></h1>
+<div class="section" id="submodules">
+<h2>Submodules<a class="headerlink" href="#submodules" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="input-mbox-keyword-clustering-module">
+<h2>input.mbox.keyword_clustering module<a class="headerlink" href="#input-mbox-keyword-clustering-module" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="input-mbox-keyword-digest-module">
+<h2>input.mbox.keyword_digest module<a class="headerlink" href="#input-mbox-keyword-digest-module" title="Permalink to this headline">¶</a></h2>
+</div>
+<div class="section" id="module-input.mbox.mbox_hdr">
+<span id="input-mbox-mbox-hdr-module"></span><h2>input.mbox.mbox_hdr module<a class="headerlink" href="#module-input.mbox.mbox_hdr" title="Permalink to this headline">¶</a></h2>
+<dl class="function">
+<dt id="input.mbox.mbox_hdr.extract_mail_header">
+<code class="descclassname">input.mbox.mbox_hdr.</code><code class="descname">extract_mail_header</code><span class="sig-paren">(</span><em>mbox_filename</em>, <em>json_filename='headers.json'</em>, <em>thread_uid_filename='thread_uid_map.json'</em>, <em>author_uid_filename='author_uid_map.json'</em><span class="sig-paren">)</span><a class="headerlink" href="#input.mbox.mbox_hdr.extract_mail_header" title="Permalink to this definition">¶</a></dt>
+<dd><p>From the .MBOX file, this function extracts the header information is extracted using two predefined classes
+available in the Python Standard Library: Mailbox and Message, for accessing and manipulating on-disk mailboxes
+and the messages they contain respectively. The headers are then saved to a JSOn file. an unique Message-ID is
+provided to each message in the .MBOX file in the ascending order of their arrival times. The The mapping between
+the UIDs and Message-IDs are stored in another JSON file
+:param mbox_filename: Contains the absolue or relative address of the file to be opened</p>
+</dd></dl>
+
+</div>
+<div class="section" id="module-input.mbox">
+<span id="module-contents"></span><h2>Module contents<a class="headerlink" href="#module-input.mbox" title="Permalink to this headline">¶</a></h2>
+</div>
+</div>
+
+
+          </div>
+        </div>
+      </div>
+      <div class="sphinxsidebar" role="navigation" aria-label="main navigation">
+        <div class="sphinxsidebarwrapper">
+  <h3><a href="index.html">Table Of Contents</a></h3>
+  <ul>
+<li><a class="reference internal" href="#">input.mbox package</a><ul>
+<li><a class="reference internal" href="#submodules">Submodules</a></li>
+<li><a class="reference internal" href="#input-mbox-keyword-clustering-module">input.mbox.keyword_clustering module</a></li>
+<li><a class="reference internal" href="#input-mbox-keyword-digest-module">input.mbox.keyword_digest module</a></li>
+<li><a class="reference internal" href="#module-input.mbox.mbox_hdr">input.mbox.mbox_hdr module</a></li>
+<li><a class="reference internal" href="#module-input.mbox">Module contents</a></li>
+</ul>
+</li>
+</ul>
+<div class="relations">
+<h3>Related Topics</h3>
+<ul>
+  <li><a href="index.html">Documentation overview</a><ul>
+  </ul></li>
+</ul>
+</div>
+  <div role="note" aria-label="source link">
+    <h3>This Page</h3>
+    <ul class="this-page-menu">
+      <li><a href="_sources/input.mbox.rst.txt"
+            rel="nofollow">Show Source</a></li>
+    </ul>
+   </div>
+<div id="searchbox" style="display: none" role="search">
+  <h3>Quick search</h3>
+    <form class="search" action="search.html" method="get">
+      <div><input type="text" name="q" /></div>
+      <div><input type="submit" value="Go" /></div>
+      <input type="hidden" name="check_keywords" value="yes" />
+      <input type="hidden" name="area" value="default" />
+    </form>
+</div>
+<script type="text/javascript">$('#searchbox').show(0);</script>
+        </div>
+      </div>
+      <div class="clearer"></div>
+    </div>
+    <div class="footer">
+      &copy;2017, Prasad Talasila.
+      
+      |
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 1.5.6</a>
+      &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.8</a>
+      
+      |
+      <a href="_sources/input.mbox.rst.txt"
+          rel="nofollow">Page source</a>
+    </div>
+
+    
+
+    
+  </body>
+</html>
