+++ docs/source/conf.py
@@ -19,7 +19,7 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('./../../lib/'))
+sys.path.insert(0, os.path.abspath('./../../../'))
 
 
 # -- General configuration ------------------------------------------------
@@ -136,7 +136,7 @@ latex_elements = {
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
     (master_doc, 'MailingListParser.tex', 'MailingListParser Documentation',
-     'Prasad Talasila, Achyudh Ram', 'manual'),
+     'Prasad, Achyudh RPrasad Talasila, Achyudh Ram', 'manual'),
 ]
