+++ requirements.txt
@@ -7,5 +7,3 @@ ipython>=3.0.0
 nltk>=3.2.2
 networkx>=1.11
 pytest>=3.0.5
-sphinx>=1.4.1
-plotly>=1.3.2
